Added ReminderViewModelItem to organize transfer of data from viewmodel to view

+++ Mindful/Models/Reminder+CoreDataClass.swift
@@ -13,10 +13,10 @@ import CoreData
 @objc(Reminder)
 public class Reminder: NSManagedObject {
 
-    func setup(_ title: String, index: Int, priority: Int, creationDate: Date) {
+    func setup(_ title: String, index: Int, priority: Int16, creationDate: Date) {
         self.title = title
         self.index = Int16(index)
-        self.priority = Int16(priority)
+        self.priority = priority
         self.creationDate = creationDate as NSDate
 
         self.completed = false
